Feature Graph

Definition of the FeatureDiGraph object used for representing directed graphs with multi-dimensional features on edges.

Index

MultiFlows.FeatureDiGraph

MultiFlows.FeatureDiGraph

MultiFlows.FeatureDiGraphEdge

Graphs.SimpleGraphs.add_edge!

Graphs.dst

Graphs.edges

Graphs.has_edge

Graphs.inneighbors

Graphs.is_directed

Graphs.ne

Graphs.nv

Graphs.outneighbors

Graphs.src

MultiFlows.feature_dim

MultiFlows.feature_matrix

MultiFlows.feature_matrix

Full docs

MultiFlows.FeatureDiGraph — Type

FeatureDiGraph{T,N}

Concrete directed graph with a feature vector for each arc.

The default constructor expects srcnodes, dstnodes to be vectors of source and destination vertices for each edge and arc_features to by a vector of numbers. The following example initialises a directed graph with three vertices and edges and a single Float64 feature on each edge.

Examplex

julia> g1 = FeatureDiGraph([1,2,3], [2,3,1], [5., 5., 5.])
FeatureDiGraph{Int64, Float64}([1, 2, 3], [2, 3, 1], [5.0, 5.0, 5.0])

MultiFlows.FeatureDiGraph — Method

FeatureDiGraph(srcnodes::Vector{T}, dstnodes::Vector{T}, arc_features::AbstractArray{N}) where {T<:Number, N<:Number}

Construct a FeatureDiGraph object where edge features are given by arc_features[i,:].

For example we can build a graph with a two-dimensional feature on each edge :

julia> g = FeatureDiGraph([1,2,3], [2,3,1], 5 * ones(3,2))
FeatureDiGraph{Int64, Vector{Float64}}([1, 2, 3], [2, 3, 1], [[5.0, 5.0], [5.0, 5.0], [5.0, 5.0]])

MultiFlows.FeatureDiGraphEdge — Type

FeatureDiGraphEdge

Concrete type representing FeatureDiGraph edges.

Graphs.SimpleGraphs.add_edge! — Method

Graphs.add_edge!(g::FeatureDiGraph{T,N}, src::T, dst::T, feat::N}

Add arc to a FeatureDiGraph object going from vertex src to dst and with features feat. Return true on success and false if graph already has an edge (src, dst).

Examples

julia> add_edge!(g1, 1, 4, 2.)
true

julia> add_edge!(g1, 1, 2, 3.)
false

julia> nv(g1), ne(g1)
(4, 4)

julia> add_edge(g, 1, 4, ones(3)) # must have feature_dim(g) == size(feat)
ERROR: DimensionMismatch("Expected feature dimension (2,) got (3,)")
[...]

Graphs.dst — Method

Graphs.dst(e::FeatureDiGraphEdge)

Get edge destination.

Graphs.edges — Method

edges(g::FeatureDiGraph)

Return list of edges of the graph. Needed for compatibility with Graphs.jl package.

Examples

julia> edges(g1)
4-element Vector{MultiFlows.FeatureDiGraphEdge{Int64, Int64}}:
 MultiFlows.FeatureDiGraphEdge{Int64, Int64}(1, 2, 1)
 MultiFlows.FeatureDiGraphEdge{Int64, Int64}(2, 3, 1)
 MultiFlows.FeatureDiGraphEdge{Int64, Int64}(3, 1, 1)
 MultiFlows.FeatureDiGraphEdge{Int64, Int64}(1, 4, 2)

Graphs.has_edge — Method

Graphs.has_edge(g::FeatureDiGraph{T,N}, s::T, d::T)

Check if graph contains edge (s,d).

Examples

julia> using Graphs

julia> has_edge(g, 1, 2)
true

julia> has_edge(g, 2, 1)
false

Graphs.inneighbors — Method

Graphs.inneighbors(g::FeatureDiGraph{T,N}, v::T)

Get neighbors u of vertex v such that edge (u,v) belongs to the graph. Needed for compatibility with Graphs.jl package.

Examples

julia> inneighbors(g1, 1)
1-element Vector{Int64}:
 3

Graphs.is_directed — Method

Graphs.is_directed(g::FeatureDiGraph{T,N})

Check if the graph is directed, always returns true. Needed for compatibility with the Graphs.jl package.

Graphs.ne — Method

ne(g::FeatureDiGraph)

Return number of arcs of the graph. Needed for compatibility with the Graphs.jl package.

Examples

julia> ne(g)
3

Graphs.nv — Method

nv(g::FeatureDiGraph)

Returns number of vertices of the graph. Needed for compatibility with the Graphs.jl package.

Examples

julia> nv(g)
3

Graphs.outneighbors — Method

Graphs.outneighbors(g::FeatureDiGraph{T,N}, v::T)

Get outgoing neighbors of vertex v in the graph. Needed for compatibility with Graphs.jl package.

Examples

julia> outneighbors(g1, 1)
2-element Vector{Int64}:
 2
 4

Graphs.src — Method

Graphs.src(e::FeatureDiGraphEdge)

Get edge source.

MultiFlows.feature_dim — Method

feature_dim(g::FeatureDiGraph)

Get arc feature dimension.

Examplex

julia> feature_dim(g1) # scalar features
()

julia> feature_dim(g)  # two-dimension features
(2,)

MultiFlows.feature_matrix — Method

feature_matrix(g::FeatureDiGraph, feature_idx::AbstractArray{Int64})

Get a nv(g) x nv(g) x size(feature_idx) matrix with coefficients equal to arc feature values corresponding to indexes in feature_idx. TODO : managing feature_idx with multiple dimensions.

Examples

julia> feature_matrix(g, [2, 1])
4×4×2 Array{Float64, 3}:
[:, :, 1] =
 0.0  5.0  0.0  1.0
 0.0  0.0  5.0  0.0
 5.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

[:, :, 2] =
 0.0  5.0  0.0  1.0
 0.0  0.0  5.0  0.0
 5.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

MultiFlows.feature_matrix — Method

feature_matrix(g::FeatureDiGraph, feature_idx::Int64)

Get a nv(g) x nv(g) matrix with coefficients equal to arc feature values. Values returned as a sparse matrix.

Examples

julia> feature_matrix(g, 1)
4×4 SparseMatrixCSC{Float64, Int64} with 4 stored entries:
  ⋅   5.0   ⋅   1.0
  ⋅    ⋅   5.0   ⋅
 5.0   ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅