Multi-Commodity Flow

MultiFlows.jl offers the MCF interface for creating Multi-Commodity Flow problems. An MCF instance is defined by a capacitated graph and a set of demands. The objective is to find a set of paths that satisfy the demands without exceeding edge capacities.

Index

MultiFlows.AbstractDemand

MultiFlows.Demand

MultiFlows.MCF

MultiFlows.MCF

Base.show

Graphs.ne

Graphs.nv

MultiFlows.capacity_matrix

MultiFlows.cost_matrix

MultiFlows.nk

MultiFlows.normalize

MultiFlows.scale

MultiFlows.scale_demands

MultiFlows.weight_matrix

Full docs

MultiFlows.AbstractDemand — Type

AbstractDemand

Abstract type subtyped for creating MCF demands.

MultiFlows.Demand — Type

MCFDemand

Concrete demand type for MCF problems. A demand has a source node src, destination node dst and an amount amount.

MultiFlows.MCF — Type

MCF

Multi-Commodity Flow problem data container. The default expects graph to be a FeatureDiGraph object with 2-dimension edge features [cost, capacity]. demands must be a list of Demand objects.

Examples

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,3], ones(4,2))
FeatureDiGraph{Int64, Vector{Float64}}([1, 2, 3, 1], [2, 3, 1, 3], [[1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0]])

julia> MCF(g, [Demand(1, 2, 1.0)])
MCF(nv = 3, ne = 4, nk = 1)
	Demand{Int64, Float64}(1, 2, 1.0)

MultiFlows.MCF — Method

MCF(g::AbstractGraph{T}, cost::Vector{N}, capacity::Vector{N}, demands::Vector{Demand{T,N})

Create a MCF object from an AbstractGraph object, a cost and capacity vector with length ne(g) and a set of demands.

Examples

julia> using Graphs

julia> gr = grid((3,2));

julia> MCF(gr, ones(ne(gr)), ones(ne(gr)), [Demand(1,6,1.0)])
MCF(nv = 6, ne = 7, nk = 1)
	Demand{Int64, Float64}(1, 6, 1.0)

cost, capacity vectors must have same length as ne(gr) :

Base.show — Method

Base.show(io::IO, pb::MCF)

Show MCF object.

Graphs.ne — Method

ne(pb::MCF)

Number of edges in the MCF network.

Examples

julia> ne(pb)
7

Graphs.nv — Method

nv(pb::MCF)

Number of vertices.

Examples

julia> nv(pb)
6

MultiFlows.capacity_matrix — Method

capacity_matrix(pb::MCF)

Return a sparse matrix with dimension (nv(pb), nv(pb)) with values equal to arc capacity.

Examples

julia> capacity_matrix(pb)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   2.0   ⋅   2.0   ⋅    ⋅
  ⋅    ⋅   2.0   ⋅   2.0   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   2.0
  ⋅    ⋅    ⋅    ⋅   2.0   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   2.0
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅

MultiFlows.cost_matrix — Method

cost_matrix(pb::MCF)

Return a sparse matrix with dimension (nv(pb), nv(pb)) with values equal to arc costs.

Examples

julia> cost_matrix(pb)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   1.0   ⋅   1.0   ⋅    ⋅
  ⋅    ⋅   1.0   ⋅   1.0   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   1.0
  ⋅    ⋅    ⋅    ⋅   1.0   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   1.0
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅

MultiFlows.nk — Method

nk(pb::MCF)

Number of demands in the MCF instance.

Examples

julia> nk(pb)
1

MultiFlows.normalize — Method

normalize(pb::MCF)

Normalize MCF instance. Costs are scaled by 1 / max(pb.cost), capacity and demand amount are scaled by 1 / max(max(pb.capacity), max(pb.amount)).

Example

julia> pb = MCF(grid((3,2)), collect(1.0:7.0), collect(0.0:2:13.0), [Demand(1,6,10.0)])
MCF(nv = 6, ne = 7, nk = 1)
	Demand{Int64, Float64}(1, 6, 10.0)

julia> pbn = normalize(pb)
MCF(nv = 6, ne = 7, nk = 1)
	Demand{Int64, Float64}(1, 6, 0.8333333333333333)

julia> cost_matrix(pbn)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   0.142857   ⋅        0.285714   ⋅         ⋅
  ⋅    ⋅        0.428571   ⋅        0.571429   ⋅
  ⋅    ⋅         ⋅         ⋅         ⋅        0.714286
  ⋅    ⋅         ⋅         ⋅        0.857143   ⋅
  ⋅    ⋅         ⋅         ⋅         ⋅        1.0
  ⋅    ⋅         ⋅         ⋅         ⋅         ⋅

julia> capacity_matrix(pbn)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   0.0   ⋅        0.166667   ⋅         ⋅
  ⋅    ⋅   0.333333   ⋅        0.5        ⋅
  ⋅    ⋅    ⋅         ⋅         ⋅        0.666667
  ⋅    ⋅    ⋅         ⋅        0.833333   ⋅
  ⋅    ⋅    ⋅         ⋅         ⋅        1.0
  ⋅    ⋅    ⋅         ⋅         ⋅         ⋅

MultiFlows.scale — Method

scale(pb::MCF, cost_factor=1.0, capacity_factor=1.0)

Return a new MCF instance with costs scaled by a cost_factor, capacity and demand amounts scaled by capacity_factor.

Example

julia> pb1 = scale(pb, 1.5, 3)
MCF(nv = 6, ne = 7, nk = 1)
	Demand{Int64, Float64}(1, 6, 3.0)

julia> cost_matrix(pb1)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   1.5   ⋅   1.5   ⋅    ⋅
  ⋅    ⋅   1.5   ⋅   1.5   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   1.5
  ⋅    ⋅    ⋅    ⋅   1.5   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   1.5
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅

julia> capacity_matrix(pb1)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   6.0   ⋅   6.0   ⋅    ⋅ 
  ⋅    ⋅   6.0   ⋅   6.0   ⋅ 
  ⋅    ⋅    ⋅    ⋅    ⋅   6.0
  ⋅    ⋅    ⋅    ⋅   6.0   ⋅ 
  ⋅    ⋅    ⋅    ⋅    ⋅   6.0
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅ 

MultiFlows.scale_demands — Method

scale_demands(pb::MCF, factor)

Scale the amount of the demands by factor. Returns new list of demands.

Example

julia> dems = scale_demands(pb, 2)
1-element Vector{Demand{Int64, Float64}}:
 Demand{Int64, Float64}(1, 6, 2.0)

MultiFlows.weight_matrix — Function

weight_matrix(pb::MCF, idx::Int64=1)

Returns a (nv(pb), nv(pb)) matrix with elements equal to edge features corresponding to idx.

Examples

julia> pb = MCF(grid((3,2)), ones(7), 2*ones(7), [Demand(1,6,1.0)]);

julia> weight_matrix(pb)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
  ⋅   1.0   ⋅   1.0   ⋅    ⋅
  ⋅    ⋅   1.0   ⋅   1.0   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   1.0
  ⋅    ⋅    ⋅    ⋅   1.0   ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   1.0
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅