Paths

Index

MultiFlows.AbstractPath

MultiFlows.VertexPath

Base.isempty

Graphs.edges

MultiFlows.edge_indices

MultiFlows.path_weight

Full docs

MultiFlows.AbstractPath — Type

AbstractPath

Abstract type for representing paths in a graph.

MultiFlows.VertexPath — Type

VertexPath

Concrete type for paths composed of a sequence of vertex indices.

Base.isempty — Method

Base.isempty(p::VertexPath)

Check if the path is empty.

Graphs.edges — Method

MultiFlows.edge_indices — Method

edge_indices(p::VertexPath, g::AbstractGraph)

Get list of edge indices corresponding to the path.

MultiFlows.path_weight — Method

path_weight(p::VertexPath, g::AbstractGraph, aggr::Function=sum)

Compute the weight of the path in graph g.