Reading/writing MCFs

Index

MultiFlows.UnknownMultiFlowFormat

MultiFlows.is_instance_dir

MultiFlows.load

MultiFlows.load_csv

MultiFlows.save

MCF file formats

Reading CSV files

MCF instances are stored as two CSV files, a link.csv file containing edge data and a service.csv file containing the demand data.

Example of a link.csv file:

|# LinkId|srcNodeId|dstNodeId|capacity |cost|latency|
|--------|---------|---------|---------|----|-------|
|1       |1        |4        |6        |2   |0      |
|2       |1        |2        |12       |3   |0      |
|3       |2        |4        |12       |3   |0      |
|4       |4        |7        |5        |8   |0      |
|5       |2        |3        |11       |4   |0      |
|6       |3        |4        |20       |8   |0      |
|7       |4        |6        |10       |3   |0      |
|8       |6        |7        |20       |3   |0      |
|9       |1        |5        |10       |80  |0      |
|10      |5        |7        |10       |20  |0      |

Example of a service.csv file:

|# DemandId|srcNodeId|dstNodeId|amount   |latency|
|----------|---------|---------|---------|-------|
|1         |1        |7        |5        |0      |
|2         |2        |6        |5        |0      |
|3         |3        |7        |5        |0      |

If dirname is the path of a directory containing both files as described in the example above we may load an MCF instance :

julia> pb = load(dirname)
MCF(nv = 7, ne = 10, nk = 3)
	Demand{Int64, Int64}(1, 7, 5)
	Demand{Int64, Int64}(2, 6, 5)
	Demand{Int64, Int64}(3, 7, 5)

julia> adjacency_matrix(pb.graph)
7×7 SparseArrays.SparseMatrixCSC{Int64, Int64} with 10 stored entries:
 ⋅  1  ⋅  1  1  ⋅  ⋅
 ⋅  ⋅  1  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅  1  1
 ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1
 ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1
 ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅

By default the edges in the link.csv file are added only in one direction. Ensuring that edges exist in both directions can be achieved by passing edge_dir = :double to the load function. Note that for each edge that is specified in only one direction within the CSV file, features will be copied.

julia> pb = load(dirname, edge_dir=:double)
MCF(nv = 7, ne = 20, nk = 3)
	Demand{Int64, Int64}(1, 7, 5)
	Demand{Int64, Int64}(2, 6, 5)
	Demand{Int64, Int64}(3, 7, 5)

julia> adjacency_matrix(pb.graph)
7×7 SparseArrays.SparseMatrixCSC{Int64, Int64} with 20 stored entries:
 ⋅  1  ⋅  1  1  ⋅  ⋅
 1  ⋅  1  1  ⋅  ⋅  ⋅
 ⋅  1  ⋅  1  ⋅  ⋅  ⋅
 1  1  1  ⋅  ⋅  1  1
 1  ⋅  ⋅  ⋅  ⋅  ⋅  1
 ⋅  ⋅  ⋅  1  ⋅  ⋅  1
 ⋅  ⋅  ⋅  1  1  1  ⋅

Saving instances

To save an MCF instance to a directory :

julia> save(pb, "path_to_instance")

This will create two files : path_to_instance/link.csv and path_to_instance/service.csv.

Full docs

MultiFlows.UnknownMultiFlowFormat — Type

UnknownMCFFormat Exception

Exception raised when trying to load instance from unknown file format.

MultiFlows.is_instance_dir — Method

is_instance_dir(dirname::String)

Check if a directory contains files link.csv, service.csv.

julia> save(pb, "test_instance")
("test_instance/link.csv", "test_instance/service.csv")

julia> is_instance_dir("test_instance")
true

MultiFlows.load — Method

load(dirname::String; format=:csv, edge_dir=:single)

Load MultiFlow problem from file. If format=:csv uses load_csv(dirname::String) (at this time CSV files are the only supported format). edge_dir can be one of the following :

:single : each edge in the input file is interpreted as a directed edge

:double : each edge in the input file is interpreted as existing in both directions with the same attributes and features

MultiFlows.load_csv — Method

load_csv(dirname::String)

Load MultiFlow instance from CSV files. Default is to search for a link.csv and service.csv file.

MultiFlows.save — Method

save(pb::MCF, dirname::String)

Save MCF instance to dirname. Will create the files <dirname>/link.csv and <dirname>/service.csv. If folder does not exist it will be created.