Multi-Commodity Flow

MultiFlows.jl offers the MCF interface for creating Multi-Commodity Flow problems. An MCF instance is defined by a capacitated graph and a set of demands. The objective is to find a set of paths that satisfy the demands without exceeding edge capacities.

Formulation

Arc-flow formulation

The problem is defined on a directed graph $G = (V, A)$ with edge costs $r_a \in \mathbb{R}^+$ and capacities $c_a \in \mathbb{R}^+$. We are given a set of demands $K$ with each demand having an origin and destination $s_k, t_k \in V$ and an amount $b_k \in \mathbb{R}^+$. Denoting by $x_a^k \in [0, 1]$ the amount of flow for demand $k$ circulating through edge $a$ the problem can be written :

\[\begin{aligned} \min\limits & \sum\limits_{k \in K} \sum\limits_{a\in A} b_k r_a x_a^k & & (1) \\ \text{s.t.} &\sum\limits_{a \in \delta^+(v)} x_a^k - \sum\limits_{a\in\delta^-(v)} x_a^k = \left\{ \begin{array}{rl} 1 & \text{if } v=s_k \\ -1 & \text{if } v=t_k \\ 0 & \text{otherwise} \end{array} \right. & \forall k\in K, v\in V \quad & (2)\\ &\sum\limits_{k\in K} b_k x_a^k \leq c_a & \forall a \in A \quad & (3)\\ &x_a^k \in [0, 1] & \forall k\in K, a\in A \quad & (4) \end{aligned}\]

$\delta^-(u), \delta^+(u)$ respectively denote the set of incomming and outgoing edges at vertex $u$. Constraint $(2)$ are flow conservation constraints and $(3)$ the edge capacity constraints. When replacing $(4)$ by integer constraints $x_a^k \in \{0, 1\}$ we refer to this problem as the Unsplittable Multi-Commodity Flow problem, demands must then by fully routed on a single path.

Path-flow formulation

For each $k \in K$ we defined $P^k$ the set of $s_k-t_k$-paths composed of edges with capacity geater than $b_k$. For each $p \in P^k$ we denote $x_p^k$ the amount of flow circulating on $p$ for demand $k$. Let $P_a^k \subset P^k$ the set of paths that pass through edge $a$. The Path-flow formulation of the problem is then :

\[\begin{aligned} \min\limits & \sum\limits_{k \in K} \sum\limits_{p\in P^k} \sum\limits_{a\in p} b_k r_a x_p^k & & (5) \\ \text{s.t.} &\sum\limits_{p\in P^k} x_p^k = 1 & \forall k\in K \quad & (6)\\ &\sum\limits_{k\in K} \sum\limits_{p\in P^k_a} b_k x_p^k \leq c_a & \forall a \in A \quad & (7)\\ &x_p^k \in [0, 1] & \forall k\in K, p\in P^k \quad & (8) \end{aligned}\]

Constraint $(6)$ ensures that the totality of $b_k$ is routed and $(7)$ that the edge capacities are not exceeded.

Index

MultiFlows.AbstractDemand

MultiFlows.Demand

MultiFlows.MCF

MultiFlows.MCF

MultiFlows.MCFSolution

Base.show

Base.show

Graphs.has_edge

Graphs.has_edge

Graphs.ne

Graphs.nv

MultiFlows.add_demand!

MultiFlows.aggregate_demands

MultiFlows.arc_flow_value

MultiFlows.available_capacity

MultiFlows.capacities

MultiFlows.capacity_matrix

MultiFlows.cost_matrix

MultiFlows.costs

MultiFlows.demand_amounts

MultiFlows.demand_endpoints

MultiFlows.demands

MultiFlows.get_path

MultiFlows.has_demand

MultiFlows.is_feasible

MultiFlows.is_solution

MultiFlows.nk

MultiFlows.normalize

MultiFlows.objective_value

MultiFlows.path_capacity

MultiFlows.paths

MultiFlows.paths

MultiFlows.paths_from_arc_flow_values

MultiFlows.scale

MultiFlows.scale_demands

MultiFlows.solution_from_arc_flow_values

MultiFlows.used_capacity

MultiFlows.weight_matrix

Full docs

MultiFlows.AbstractDemand — Type

AbstractDemand

Abstract type subtyped for creating MCF demands.

MultiFlows.Demand — Type

MCFDemand

Concrete demand type for MCF problems. A demand has a source node src, destination node dst and an amount amount.

MultiFlows.MCF — Type

MCF

Multi-Commodity Flow problem data container. The default expects graph to be a FeatureDiGraph object with 2-dimension edge features [cost, capacity]. demands must be a list of Demand objects.

Examples

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,3], ones(4,2))
FeatureDiGraph{Int64, Vector{Float64}}([1, 2, 3, 1], [2, 3, 1, 3], [[1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0]])

julia> MCF(g, [Demand(1, 2, 1.0)])
MCF(nv = 3, ne = 4, nk = 1)
	Demand{Int64, Float64}(1, 2, 1.0)

MultiFlows.MCF — Method

MCF(g::AbstractGraph{T}, cost::AbstractVector{N}, capacity::AbstractVector{N}, demands::Vector{Demand{T,N})

Create a MCF object from an AbstractGraph object, a cost and capacity vector with length ne(g) and a set of demands.

Examples

julia> using Graphs

julia> gr = grid((3,2));

julia> MCF(gr, ones(ne(gr)), ones(ne(gr)), [Demand(1,6,1.0)])
MCF(nv = 6, ne = 14, nk = 1)
	Demand{Int64, Float64}(1, 6, 1.0)

cost, capacity vectors must have same length as ne(gr) :

Base.show — Method

Base.show(io::IO, pb::MCF)

Show MCF object.

Graphs.ne — Method

ne(pb::MCF)

Number of edges in the MCF network.

Examples

julia> ne(pb)
14

Graphs.nv — Method

nv(pb::MCF)

Number of vertices.

Examples

julia> nv(pb)
6

MultiFlows.add_demand! — Method

add_demand!(pb::MCF{T,N}, s::T, t::T, a::N) where {T,N}
add_demand!(pb::MCF{T,N}, demand::Demand{N}) where {T,N}

Add a demand to the problem.

Example

julia> pb = MCF(grid((3,2)), ones(7), ones(7), Demand{Int64,Float64}[])
MCF(nv = 6, ne = 14, nk = 0)

julia> add_demand!(pb, 1, 2, 1.0)
1-element Vector{Demand{Int64, Float64}}:
 Demand{Int64, Float64}(1, 2, 1.0)

julia> add_demand!(pb, Demand(4, 2, 1.0))
2-element Vector{Demand{Int64, Float64}}:
 Demand{Int64, Float64}(1, 2, 1.0)
 Demand{Int64, Float64}(4, 2, 1.0)

julia> pb
MCF(nv = 6, ne = 14, nk = 2)
	Demand{Int64, Float64}(1, 2, 1.0)
	Demand{Int64, Float64}(4, 2, 1.0)

MultiFlows.aggregate_demands — Method

aggregate_demands(pb::MCF)

Create new instance with aggregated demands : if demands $k_1, k_2$ are such that $s_{k_1} = s_{k_2}$ and $t_{k_1} = t_{k_2}$ they are removed from the problem and a new demand $k'$ with $b_{k'} = b_{k_1} + b_{k_2}$ and same endpoints is added.

Example

julia> pb = MCF(grid((3,2)), ones(7), ones(7), [Demand(1,3,1.), Demand(1,3,1.)])
MCF(nv = 6, ne = 14, nk = 2)
	Demand{Int64, Float64}(1, 3, 1.0)
	Demand{Int64, Float64}(1, 3, 1.0)

julia> aggregate_demands(pb)
MCF(nv = 6, ne = 14, nk = 1)
	Demand{Int64, Float64}(1, 3, 2.0)

MultiFlows.capacities — Method

capacities(pb::MCF)

Return edge capacities in a Vector.

Example

julia> x, y = rand(14), rand(14);

julia> pb = MCF(grid((3,2)), x, y, [Demand(1,2,1.)]);

julia> capacities(pb) == y
true

MultiFlows.capacity_matrix — Method

capacity_matrix(pb::MCF)

Return a sparse matrix with dimension (nv(pb), nv(pb)) with values equal to arc capacity.

Examples

julia> capacity_matrix(pb)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅   2.0   ⋅   2.0   ⋅    ⋅
 2.0   ⋅   2.0   ⋅   2.0   ⋅
  ⋅   2.0   ⋅    ⋅    ⋅   2.0
 2.0   ⋅    ⋅    ⋅   2.0   ⋅
  ⋅   2.0   ⋅   2.0   ⋅   2.0
  ⋅    ⋅   2.0   ⋅   2.0   ⋅

MultiFlows.cost_matrix — Method

cost_matrix(pb::MCF)

Return a sparse matrix with dimension (nv(pb), nv(pb)) with values equal to arc costs.

Examples

julia> cost_matrix(pb)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅   1.0   ⋅   1.0   ⋅    ⋅ 
 1.0   ⋅   1.0   ⋅   1.0   ⋅ 
  ⋅   1.0   ⋅    ⋅    ⋅   1.0
 1.0   ⋅    ⋅    ⋅   1.0   ⋅ 
  ⋅   1.0   ⋅   1.0   ⋅   1.0
  ⋅    ⋅   1.0   ⋅   1.0   ⋅ 

MultiFlows.costs — Method

costs(pb::MCF)

Return edge costs in a Vector.

Example

julia> x, y = rand(14), rand(14);

julia> pb = MCF(grid((3,2)), x, y, [Demand(1,2,1.)]);

julia> costs(pb) == x
true

MultiFlows.demand_amounts — Method

demand_amounts(pb::MCF)

Return a vector of the demand amounts.

Example

julia> pb = MCF(grid((3,2)), ones(7), ones(7), [Demand(1,2,1.), Demand(1,3,.5)]);

julia> demand_amounts(pb)
2-element Vector{Float64}:
 1.0
 0.5

MultiFlows.demand_endpoints — Method

demand_endpoints(pb::MCF)

Return vector of source and destination nodes for each demand.

Example

julia> pb = MCF(grid((3,2)), ones(7), ones(7), [Demand(1,2,1.), Demand(1,3,.5)]);

julia> demand_endpoints(pb)
([1, 1], [2, 3])

MultiFlows.demands — Method

demands(pb::MCF{T,N}, s::T, t::T)

Get demands with origin and destination s,t.

Example

julia> demands(pb, 1, 2)
1-element Vector{Demand{Int64, Float64}}:
 Demand{Int64, Float64}(1, 2, 1.0)

julia> demands(pb, 2, 1)
Demand{Int64, Float64}[]

MultiFlows.get_path — Method

get_path(pb::MCF{T,N}, s::T, t::T; 
         alg::Function=(s,t)->enumerate_paths(dijkstra_shortest_paths(pb.graph, s), t)
        ) where {T,N}

Get path from s to t for MCF problem pb.

Example

julia> get_path(pb, 1, 6)
VertexPath{Int64}([1, 2, 5, 6])

MultiFlows.has_demand — Method

has_demand(pb::MCF{T,N}, s::T, d::T)

Check if problem has a demand originating at vertex s and with destination d.

Example

julia> pb = MCF(grid((3,2)), ones(7), ones(7), [Demand(1,2,1.)]);

julia> has_demand(pb, 1, 2)
true

julia> has_demand(pb, 2, 1)
false

MultiFlows.nk — Method

nk(pb::MCF)

Number of demands in the MCF instance.

Examples

julia> nk(pb)
1

MultiFlows.normalize — Method

normalize(pb::MCF)

Normalize MCF instance. Costs are scaled by 1 / max(pb.cost), capacity and demand amount are scaled by 1 / max(max(pb.capacity), max(pb.amount)).

Example

julia> pb = MCF(grid((3,2)), collect(1.0:7.0), collect(0.0:2:13.0), [Demand(1,6,10.0)])
MCF(nv = 6, ne = 14, nk = 1)
	Demand{Int64, Float64}(1, 6, 10.0)

julia> pbn = normalize(pb)
MCF(nv = 6, ne = 14, nk = 1)
	Demand{Int64, Float64}(1, 6, 0.8333333333333333)

julia> cost_matrix(pbn)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅        0.142857   ⋅        0.285714   ⋅         ⋅ 
 0.142857   ⋅        0.428571   ⋅        0.571429   ⋅ 
  ⋅        0.428571   ⋅         ⋅         ⋅        0.714286
 0.285714   ⋅         ⋅         ⋅        0.857143   ⋅ 
  ⋅        0.571429   ⋅        0.857143   ⋅        1.0
  ⋅         ⋅        0.714286   ⋅        1.0        ⋅ 

julia> capacity_matrix(pbn)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅        0.0        ⋅        0.166667   ⋅         ⋅ 
 0.0        ⋅        0.333333   ⋅        0.5        ⋅ 
  ⋅        0.333333   ⋅         ⋅         ⋅        0.666667
 0.166667   ⋅         ⋅         ⋅        0.833333   ⋅ 
  ⋅        0.5        ⋅        0.833333   ⋅        1.0
  ⋅         ⋅        0.666667   ⋅        1.0        ⋅ 

MultiFlows.scale — Method

scale(pb::MCF, cost_factor=1.0, capacity_factor=1.0)

Return a new MCF instance with costs scaled by a cost_factor, capacity and demand amounts scaled by capacity_factor.

Example

julia> pb1 = scale(pb, 1.5, 3)
MCF(nv = 6, ne = 14, nk = 1)
	Demand{Int64, Float64}(1, 6, 3.0)

julia> cost_matrix(pb1)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅   1.5   ⋅   1.5   ⋅    ⋅
 1.5   ⋅   1.5   ⋅   1.5   ⋅
  ⋅   1.5   ⋅    ⋅    ⋅   1.5
 1.5   ⋅    ⋅    ⋅   1.5   ⋅
  ⋅   1.5   ⋅   1.5   ⋅   1.5
  ⋅    ⋅   1.5   ⋅   1.5   ⋅

julia> capacity_matrix(pb1)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅   6.0   ⋅   6.0   ⋅    ⋅
 6.0   ⋅   6.0   ⋅   6.0   ⋅
  ⋅   6.0   ⋅    ⋅    ⋅   6.0
 6.0   ⋅    ⋅    ⋅   6.0   ⋅
  ⋅   6.0   ⋅   6.0   ⋅   6.0
  ⋅    ⋅   6.0   ⋅   6.0   ⋅

MultiFlows.scale_demands — Method

scale_demands(pb::MCF, factor)

Scale the amount of the demands by factor. Returns new list of demands.

Example

julia> dems = scale_demands(pb, 2)
1-element Vector{Demand{Int64, Float64}}:
 Demand{Int64, Float64}(1, 6, 2.0)

MultiFlows.weight_matrix — Function

weight_matrix(pb::MCF, idx::Int64=1)

Returns a (nv(pb), nv(pb)) matrix with elements equal to edge features corresponding to idx.

Examples

julia> pb = MCF(grid((3,2)), ones(7), 2*ones(7), [Demand(1,6,1.0)]);

julia> weight_matrix(pb)
6×6 SparseArrays.SparseMatrixCSC{Float64, Int64} with 14 stored entries:
  ⋅   1.0   ⋅   1.0   ⋅    ⋅
 1.0   ⋅   1.0   ⋅   1.0   ⋅
  ⋅   1.0   ⋅    ⋅    ⋅   1.0
 1.0   ⋅    ⋅    ⋅   1.0   ⋅
  ⋅   1.0   ⋅   1.0   ⋅   1.0
  ⋅    ⋅   1.0   ⋅   1.0   ⋅

MultiFlows.MCFSolution — Type

MCFSolution

Container representing the solution of an MCF problem. Contains a list of paths such that sol.paths[k] contains the paths used to route demand k and flows[k] are the flow values

Base.show — Method

Base.show(sol::MCFSolution)

Show solutions.

Graphs.has_edge — Method

Graphs.has_edge(sol::MCFSolution, k::Int64, s::T, t::T) where {T}

Check if the solution uses edge (s,t) for demand k.

Example

julia> sol = MCFSolution([[VertexPath([1,2])], [VertexPath([1,4,5,2])]], [[.5], [.5]]);

julia> has_edge(sol, 1, 1, 2)
true

julia> has_edge(sol, 1, 5, 4)
false

Graphs.has_edge — Method

Graphs.has_edge(sol::MCFSolution, s::T, t::T) where {T}

Check if the solution uses edge (s,t).

Example

julia> sol = MCFSolution([[VertexPath([1,2])], [VertexPath([1,4,5,2])]], [[.5], [.5]]);

julia> has_edge(sol, 1, 2)
true

julia> has_edge(sol, 5, 4)
false

MultiFlows.arc_flow_value — Method

arc_flow_value(sol::MCFSolution, pb::MCF)

Returns a (nk(pb), ne(pb)) sized matrix where coefficients x[k,a] is the amount of flow for demand k circulating through a.

Example

julia> pb = MCF(grid((3,2)), collect(1:7.), collect(10:16.), [Demand(1,2,1.)]);

julia> sol = MCFSolution([[VertexPath([1,2]), VertexPath([1,4,5,2])]], [[.5, .5]]);

julia> arc_flow_value(sol, pb)
1×14 Matrix{Float64}:
 0.5  0.5  0.0  0.0  0.0  0.5  0.0  0.0  0.0  0.0  0.5  0.0  0.0  0.0

MultiFlows.available_capacity — Method

available_capacity(sol::MCFSolution, pb::MCF)

Compute available capacities on the graph for a given solution.

Example

julia> pb = MCF(grid((3,2)), ones(14), ones(14), [Demand(1,2,1.)]);

julia> sol = MCFSolution([[VertexPath([1,4,5,2])]], [[1.]]);

julia> available_capacity(sol, pb)
14-element Vector{Float64}:
 1.0
 0.0
 1.0
 1.0
 1.0
 0.0
 1.0
 1.0
 1.0
 1.0
 0.0
 1.0
 1.0
 1.0

MultiFlows.is_feasible — Method

is_feasible(sol::MCFSolution, pb::MCF)

Check if the solution is feasible, has to be a valid solution for the problem and the total amount circulating on the graph must not be greater than the edge capacities and all demands must be routed entirely.

Example

julia> pb = MCF(grid((3,2)), ones(14), ones(14), [Demand(1,2,2.)]);

julia> sol = MCFSolution([[VertexPath([1,4,5,2])]], [[1.]]);

julia> is_feasible(sol, pb)
false

julia> sol = MCFSolution([[VertexPath([1,2]), VertexPath([1,4,5,2])]], [[.5, .5]]);

julia> is_feasible(sol, pb)
true

MultiFlows.is_solution — Method

is_solution(sol::MCFSolution, pb::MCF)

Check if sol is a solution for problem pb. Checks if length(paths) == length(flows) == nk(pb), that each path is a valid path on the graph and has the correct endpoints, and that sum(flows[k]) <= 1.

Example

julia> pb = MCF(grid((3,2)), ones(7), ones(7), [Demand(1,2,1.)])
MCF(nv = 6, ne = 14, nk = 1)
	Demand{Int64, Float64}(1, 2, 1.0)

julia> sol = MCFSolution([[VertexPath([1,2])]], [[1.]])
MCFSolution
	Demand k = 1
		1.0 on VertexPath{Int64}([1, 2])

julia> is_solution(sol, pb)
true

julia> sol = MCFSolution([[VertexPath([1,2]), VertexPath([1,4,5,2])]], [[.5, .5]])
MCFSolution
	Demand k = 1
		0.5 on VertexPath{Int64}([1, 2])
		0.5 on VertexPath{Int64}([1, 4, 5, 2])

julia> is_solution(sol, pb)
true

Lets create invalid solutions :

julia> # path is not valid

julia> is_solution(MCFSolution([[VertexPath([1,3])]], [[1.]]), pb)
false

julia> # total flow for demand 1 is greater than 1

julia> is_solution(MCFSolution([[VertexPath([1,2])]], [[2.]]), pb)
false

MultiFlows.objective_value — Method

objective_value(sol::MCFSolution, pb::MCF)

Compute the objective value of sol.

Example

julia> pb = MCF(grid((3,2)), ones(14), ones(14), [Demand(1,2,2.)]);

julia> sol = MCFSolution([[VertexPath([1,4,5,2])]], [[1.]]);

julia> objective_value(sol, pb)
6.0

julia> sol = MCFSolution([[VertexPath([1,2]), VertexPath([1,4,5,2])]], [[.5, .5]]);

julia> objective_value(sol, pb)
4.0

MultiFlows.path_capacity — Method

path_capacity(p::VertexPath, pb::MCF)

Get minimum capacity of arcs belonging to path p.

Example

julia> pb = MCF(grid((3,2)), ones(Int64,7), 1:7, [Demand(1,2,2)]);

julia> p = get_path(pb, 1, 6)
VertexPath{Int64}([1, 2, 5, 6])

julia> path_capacity(p, pb)
1

MultiFlows.paths — Method

paths(sol::MCFSolution, k::Int64)

Return paths used in the solution with non-zero flow for demand k.

Example

julia> sol = MCFSolution([[VertexPath([1,2])], [VertexPath([1,4,5,2])]], [[.5], [.5]]);

julia> paths(sol, 1)
1-element Vector{VertexPath{Int64}}:
 VertexPath{Int64}([1, 2])

MultiFlows.paths — Method

paths(sol::MCFSolution)

Return all paths used in the solution with non-zero flow.

Example

julia> sol = MCFSolution([[VertexPath([1,2]), VertexPath([1,4,5,2])]], [[.5, .5]]);

julia> paths(sol)
2-element Vector{VertexPath{Int64}}:
 VertexPath{Int64}([1, 2])
 VertexPath{Int64}([1, 4, 5, 2])

MultiFlows.paths_from_arc_flow_values — Method

paths_from_arc_flow_values(x::Vector, k::Int64, pb::MCF)

Compute paths and flows from arc flow values.

Example

julia> pb = MCF(grid((3,2)), ones(14), ones(14), [Demand(1,2,2.)]);

julia> sol = MCFSolution([[VertexPath([1,4,5,2])]], [[1.]]);

julia> x = arc_flow_value(sol, pb)
1×14 Matrix{Float64}:
 0.0  1.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0

julia> paths_from_arc_flow_values(x[1,:], 1, pb)
(VertexPath[VertexPath{Int64}([1, 4, 5, 2])], [1.0])

MultiFlows.solution_from_arc_flow_values — Method

from_arc_flow_values(x::AbstractMatrix{Float64}, pb::MCF)

Compute solution paths from $x_a^k$ values.

Example

julia> pb = MCF(grid((3,2)), ones(14), ones(14), [Demand(1,2,2.)]);

julia> sol = MCFSolution([[VertexPath([1,4,5,2])]], [[1.]]);

julia> x = arc_flow_value(sol, pb)
1×14 Matrix{Float64}:
 0.0  1.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0

julia> solution_from_arc_flow_values(x, pb)
MCFSolution
	Demand k = 1
		1.0 on VertexPath{Int64}([1, 4, 5, 2])

MultiFlows.used_capacity — Method

used_capacity(sol::MCFSolution, pb::MCF)

Compute total edge capacity used by the solution.

Example

julia> pb = MCF(grid((3,2)), ones(14), ones(14), [Demand(1,2,1.)]);

julia> sol = MCFSolution([[VertexPath([1,4,5,2])]], [[1.]]);

julia> used_capacity(sol, pb)
14-element Vector{Float64}:
 0.0
 1.0
 0.0
 0.0
 0.0
 1.0
 0.0
 0.0
 0.0
 0.0
 1.0
 0.0
 0.0
 0.0