Feature Graph

Definition of the FeatureDiGraph object used for representing directed graphs with multi-dimensional features on edges.

Index

MultiFlows.FeatureDiGraph

MultiFlows.FeatureDiGraphEdge

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.add_edge!

Graphs.dst

Graphs.edges

Graphs.has_edge

Graphs.inneighbors

Graphs.is_directed

Graphs.ne

Graphs.nv

Graphs.outneighbors

Graphs.src

MultiFlows.feature_dim

MultiFlows.feature_matrix

MultiFlows.feature_matrix

Full docs

MultiFlows.FeatureDiGraph — Type

FeatureDiGraph

Concrete directed graph with a feature vector for each arc.

MultiFlows.FeatureDiGraphEdge — Type

FeatureDiGraphEdge

Concrete type representing FeatureDiGraph edges.

Graphs.SimpleGraphs.add_edge! — Method

Graphs.add_edge!(g::FeatureDiGraph{T,N}, src, dst, feat::AbstractArray{N}}

Add arc to a FeatureDiGraph object going from vertex src to dst and with features feat.

Graphs.SimpleGraphs.add_edge! — Method

Graphs.add_edge!(g::FeatureDiGraph{T,N}, src, dst, feat}

Add arc to a FeatureDiGraph object going from vertex src to dst and with features feat.

Graphs.dst — Method

Graphs.dst(e::FeatureDiGraphEdge)

Get edge destination.

Graphs.edges — Method

edges(g::FeatureDiGraph)

Return list of edges of the graph.

Graphs.has_edge — Method

Graphs.has_edge(g::FeatureDiGraph{T,N}, s::T, d::T)

Check if graph contains edge (s,d)

Graphs.inneighbors — Method

Graphs.inneighbors(g::FeatureDiGraph{T,N}, v::T)

Get neighbors u of vertex v such that edge (u,v) belongs to the graph.

Graphs.is_directed — Method

Graphs.is_directed(g::FeatureDiGraph{T,N})

Always returns true.

Graphs.ne — Method

ne(g::FeatureDiGraph)

Return number of arcs of the graph.

Graphs.nv — Method

nv(g::FeatureDiGraph)

Returns number of vertices of the graph.

Graphs.outneighbors — Method

Graphs.outneighbors(g::FeatureDiGraph{T,N}, v::T)

Get outgoing neighbors of vertex v in the graph.

Graphs.src — Method

Graphs.src(e::FeatureDiGraphEdge)

Get edge source.

MultiFlows.feature_dim — Method

feature_dim(g::FeatureDiGraph)

Get arc feature dimension.

MultiFlows.feature_matrix — Method

feature_matrix(g::FeatureDiGraph, feature_idx::AbstractArray{Int64})

Get a (nv(g) x nv(g) x size(featureidx)) matrix with coefficients equal to arc feature values. TODO : managing featureidx with multiple dimensions.

MultiFlows.feature_matrix — Method

feature_matrix(g::FeatureDiGraph, feature_idx::Int64)

Get a (nv(g) x nv(g)) matrix with coefficients equal to arc feature values.