Feature Graph

Definition of the FeatureDiGraph object used for representing directed graphs with multi-dimensional features on edges.

Index

MultiFlows.FeatureDiGraph

MultiFlows.FeatureDiGraph

MultiFlows.FeatureDiGraph

MultiFlows.FeatureDiGraphEdge

Base.reverse

Graphs.SimpleGraphs.add_edge!

Graphs.SimpleGraphs.rem_edge!

Graphs.dst

Graphs.edges

Graphs.has_edge

Graphs.inneighbors

Graphs.is_directed

Graphs.ne

Graphs.nv

Graphs.outneighbors

Graphs.src

Graphs.weights

MultiFlows.convert_features

MultiFlows.double_edges!

MultiFlows.edge_features

MultiFlows.edge_features

MultiFlows.edge_index_matrix

MultiFlows.feature_dim

MultiFlows.feature_matrix

MultiFlows.feature_matrix

MultiFlows.scale_features

MultiFlows.zero_features

Full docs

MultiFlows.FeatureDiGraph — Type

FeatureDiGraph{T,N}

Concrete directed graph with a feature vector for each edge.

The default constructor expects srcnodes, dstnodes to be vectors of source and destination vertices for each edge and features to by a vector of numbers. The following example initialises a directed graph with three vertices and edges and a single Float64 feature on each edge.

Examplex

julia> using MultiFlows

julia> g1 = FeatureDiGraph([1,2,3], [2,3,1], [5., 5., 5.])
FeatureDiGraph{Int64, Float64}([1, 2, 3], [2, 3, 1], [5.0, 5.0, 5.0])

MultiFlows.FeatureDiGraph — Method

FeatureDiGraph(g::AbstractGraph{T}, features::AbstractArray{N})

Construct a feature graph from an AbstractGraph object and a set of features. If g is an undirected graph the edges will be added in both directions and features should have first dimension either ne(g) in which case features are duplicated in both directions, or 2*ne(g).

Examples

julia> using Graphs

julia> gr = complete_digraph(5)
{5, 20} directed simple Int64 graph

julia> FeatureDiGraph(gr, ones(ne(gr), 2))
FeatureDiGraph{Int64, Vector{Float64}}([1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4, 5, 5, 5, 5], [2, 3, 4, 5, 1, 3, 4, 5, 1, 2, 4, 5, 1, 2, 3, 5, 1, 2, 3, 4], [[1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0]])

julia> gr = complete_graph(5)
{5, 10} undirected simple Int64 graph

julia> FeatureDiGraph(gr, ones(2*ne(gr), 2))
FeatureDiGraph{Int64, Vector{Float64}}([1, 1, 1, 1, 2, 2, 2, 3, 3, 4, 2, 3, 4, 5, 3, 4, 5, 4, 5, 5], [2, 3, 4, 5, 3, 4, 5, 4, 5, 5, 1, 1, 1, 1, 2, 2, 2, 3, 3, 4], [[1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0], [1.0, 1.0]])

MultiFlows.FeatureDiGraph — Method

FeatureDiGraph(srcnodes::Vector{T}, dstnodes::Vector{T}, features::AbstractArray{N}) where {T<:Number, N<:Number}

Construct a FeatureDiGraph object where edge features are given by features[i,:].

Examplex

For example we can build a graph with a two-dimensional feature on each edge :

julia> g = FeatureDiGraph([1,2,3], [2,3,1], 5 * ones(3,2))
FeatureDiGraph{Int64, Vector{Float64}}([1, 2, 3], [2, 3, 1], [[5.0, 5.0], [5.0, 5.0], [5.0, 5.0]])

MultiFlows.FeatureDiGraphEdge — Type

FeatureDiGraphEdge

Concrete type representing FeatureDiGraph edges.

Base.reverse — Method

Graphs.SimpleGraphs.add_edge! — Method

Graphs.add_edge!(g::FeatureDiGraph{T,N}, src::T, dst::T, feat::N=zero_feature(g)}

Add edge to a FeatureDiGraph object going from vertex src to dst and with features feat. Return true on success and false if graph already has an edge (src, dst). If force=true the edge is added even if it already exists.

Examples

julia> g = FeatureDiGraph([1,2,3], [2,3,1], ones(3,2));

julia> add_edge!(g, 1, 4, [2., 2.])
true

julia> add_edge!(g, 1, 2, [3., 3.])
false

julia> nv(g), ne(g)
(4, 4)

julia> add_edge!(g, 1, 4, ones(3)) # must have feature_dim(g) == size(feat)
ERROR: DimensionMismatch("Expected feature dimension (2,) got (3,)")
[...]

Graphs.SimpleGraphs.rem_edge! — Method

Graphs.dst — Method

Graphs.dst(e::FeatureDiGraphEdge)

Get edge destination.

Graphs.edges — Method

edges(g::FeatureDiGraph)

Return list of edges of the graph. Needed for compatibility with Graphs.jl package.

Examples

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], [1,1,1,2]);

julia> edges(g)
4-element Vector{FeatureDiGraphEdge{Int64, Int64}}:
 FeatureDiGraphEdge{Int64, Int64}(1, 2, 1)
 FeatureDiGraphEdge{Int64, Int64}(2, 3, 1)
 FeatureDiGraphEdge{Int64, Int64}(3, 1, 1)
 FeatureDiGraphEdge{Int64, Int64}(1, 4, 2)

Graphs.has_edge — Method

Graphs.has_edge(g::FeatureDiGraph{T,N}, s::T, d::T)

Check if graph contains edge (s,d).

Examples

julia> using Graphs

julia> has_edge(g, 1, 2)
true

julia> has_edge(g, 2, 1)
false

Graphs.inneighbors — Method

Graphs.inneighbors(g::FeatureDiGraph{T,N}, v::T)

Get neighbors u of vertex v such that edge (u,v) belongs to the graph. Needed for compatibility with Graphs.jl package.

Examples

julia> inneighbors(g, 1)
1-element Vector{Int64}:
 3

Graphs.is_directed — Method

Graphs.is_directed(g::FeatureDiGraph{T,N})

Check if the graph is directed, always returns true. Needed for compatibility with the Graphs.jl package.

Graphs.ne — Method

ne(g::FeatureDiGraph)

Return number of edge of the graph. Needed for compatibility with the Graphs.jl package.

Examples

julia> ne(g)
3

Graphs.nv — Method

nv(g::FeatureDiGraph)

Returns number of vertices of the graph. Needed for compatibility with the Graphs.jl package.

Examples

julia> nv(g)
3

Graphs.outneighbors — Method

Graphs.outneighbors(g::FeatureDiGraph{T,N}, v::T)

Get outgoing neighbors of vertex v in the graph. Needed for compatibility with Graphs.jl package.

Examples

julia> outneighbors(g, 1)
2-element Vector{Int64}:
 2
 4

Graphs.src — Method

Graphs.src(e::FeatureDiGraphEdge)

Get edge source.

Graphs.weights — Function

Graphs.weights(g::FeatureDiGraph, feature_idx::Int64=1)

Get graph weight matrix corresponding to in feature with index idx. Needed for compatibility with Graphs.jl.

Examples

julia> using Graphs

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,3],[[1,1],[1,1],[1,1],[4,1]])
FeatureDiGraph{Int64, Vector{Int64}}([1, 2, 3, 1], [2, 3, 1, 3], [[1, 1], [1, 1], [1, 1], [4, 1]])

julia> ds = dijkstra_shortest_paths(g, 1);

julia> enumerate_paths(ds, 3)
3-element Vector{Int64}:
 1
 2
 3

julia> ds = dijkstra_shortest_paths(g, 1, weights(g, 2));

julia> enumerate_paths(ds, 3)
2-element Vector{Int64}:
 1
 3

MultiFlows.convert_features — Method

convert_features(g::FeatureDiGraph{T,N}, dtype::DataType) where {T,N}

Convert graph features to type dtype.

Example

julia> pb = MCF(grid((3,2)), ones(Int64,7), 1:7, [Demand(1,2,2)]);

julia> pb.graph
FeatureDiGraph{Int64, Vector{Int64}}([1, 1, 2, 2, 3, 4, 5, 2, 4, 3, 5, 6, 5, 6], [2, 4, 3, 5, 6, 5, 6, 1, 1, 2, 2, 3, 4, 5], [[1, 1], [1, 2], [1, 3], [1, 4], [1, 5], [1, 6], [1, 7], [1, 1], [1, 2], [1, 3], [1, 4], [1, 5], [1, 6], [1, 7]])

julia> convert_features(pb.graph, Float64)
FeatureDiGraph{Int64, Vector{Float64}}([1, 1, 2, 2, 3, 4, 5, 2, 4, 3, 5, 6, 5, 6], [2, 4, 3, 5, 6, 5, 6, 1, 1, 2, 2, 3, 4, 5], [[1.0, 1.0], [1.0, 2.0], [1.0, 3.0], [1.0, 4.0], [1.0, 5.0], [1.0, 6.0], [1.0, 7.0], [1.0, 1.0], [1.0, 2.0], [1.0, 3.0], [1.0, 4.0], [1.0, 5.0], [1.0, 6.0], [1.0, 7.0]])

MultiFlows.double_edges! — Method

double_edges!(g::FeatureDiGraph)

For each edge (u,v) add edge (v,u) if not already present in the graph. Features of the edge are copied.

Examples

julia> using Graphs

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,3],[[1,1],[1,1],[1,1],[4,1]])
FeatureDiGraph{Int64, Vector{Int64}}([1, 2, 3, 1], [2, 3, 1, 3], [[1, 1], [1, 1], [1, 1], [4, 1]])

julia> ne(g)
4

julia> double_edges!(g)
FeatureDiGraph{Int64, Vector{Int64}}([1, 2, 3, 1, 2, 3], [2, 3, 1, 3, 1, 2], [[1, 1], [1, 1], [1, 1], [4, 1], [1, 1], [1, 1]])

julia> ne(g)
6

MultiFlows.edge_features — Method

edge_features(g::FeatureDiGraph, idx::Int64)

Get array of edge features corresponding to index idx. Returns a (ne(g), 1) array.

Examples

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], hcat(3*ones(4), 4*ones(4)));

julia> edge_features(g, 1)
4-element Vector{Float64}:
 3.0
 3.0
 3.0
 3.0

MultiFlows.edge_features — Method

edge_features(g::FeatureDiGraph)

Get array of edge features. Returns a (ne(g), feature_dim(g)) array.

Examples

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], hcat(3*ones(4), 4*ones(4)));

julia> edge_features(g)
4×2 Matrix{Float64}:
 3.0  4.0
 3.0  4.0
 3.0  4.0
 3.0  4.0

MultiFlows.edge_index_matrix — Method

edge_index_matrix(g::AbstractGraph)

Get (nv(g), nv(g)) sized matrix where M[i,j] is the index of edge i,j.

Examples

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], hcat(3*ones(4), 4*ones(4)));

julia> edge_index_matrix(g)
3×4 SparseArrays.SparseMatrixCSC{Int64, Int64} with 4 stored entries:
 ⋅  1  ⋅  4
 ⋅  ⋅  2  ⋅
 3  ⋅  ⋅  ⋅

Simple undirected graphs :

julia> using Graphs

julia> g = grid((3,2))
{6, 7} undirected simple Int64 graph

julia> edge_index_matrix(g)
6×6 SparseArrays.SparseMatrixCSC{Int64, Int64} with 14 stored entries:
 ⋅  1  ⋅  2  ⋅  ⋅
 1  ⋅  3  ⋅  4  ⋅
 ⋅  3  ⋅  ⋅  ⋅  5
 2  ⋅  ⋅  ⋅  6  ⋅
 ⋅  4  ⋅  6  ⋅  7
 ⋅  ⋅  5  ⋅  7  ⋅

MultiFlows.feature_dim — Method

feature_dim(g::FeatureDiGraph)

Get edge feature dimension.

Examplex

julia> g1 = FeatureDiGraph([1,2,3], [2,3,1], ones(3));

julia> feature_dim(g1) # scalar features
()

julia> g2 = FeatureDiGraph([1,2,3], [2,3,1], ones(3, 2));

julia> feature_dim(g2)  # two-dimension features
(2,)

MultiFlows.feature_matrix — Method

feature_matrix(g::FeatureDiGraph, feature_idx::AbstractArray{Int64})

Get a nv(g) x nv(g) x size(feature_idx) matrix with coefficients equal to edge feature values corresponding to indexes in feature_idx. TODO : managing feature_idx with multiple dimensions.

Examples

julia> feature_matrix(g, [2, 1])
4×4×2 Array{Float64, 3}:
[:, :, 1] =
 0.0  4.0  0.0  4.0
 0.0  0.0  4.0  0.0
 4.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

[:, :, 2] =
 0.0  3.0  0.0  3.0
 0.0  0.0  3.0  0.0
 3.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

MultiFlows.feature_matrix — Method

feature_matrix(g::FeatureDiGraph, feature_idx::Int64)

Get a nv(g) x nv(g) matrix with coefficients equal to edge feature values. Values returned as a sparse matrix.

Examples

Graph with scalar features :

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], [5.0, 5.0, 5.0, 1.0]);

julia> feature_matrix(g)
4×4 SparseArrays.SparseMatrixCSC{Float64, Int64} with 4 stored entries:
  ⋅   5.0   ⋅   1.0
  ⋅    ⋅   5.0   ⋅
 5.0   ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅

Graph with multi-dimensional features :

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], hcat(3*ones(4), 4*ones(4)));

julia> feature_matrix(g, 2)
4×4 SparseArrays.SparseMatrixCSC{Float64, Int64} with 4 stored entries:
  ⋅   4.0   ⋅   4.0
  ⋅    ⋅   4.0   ⋅
 4.0   ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅

MultiFlows.scale_features — Method

scale_features(g::FeatureDiGraph{T,N}, factor::N)

Scale the features of the graph by factor. factor should have the same dimension as the edge features.

MultiFlows.zero_features — Method

zero_features(g::FeatureDiGraph{T,N}) where {T,N}

Create zero valued features for graph g.

Example

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], [1,2,3,4]);

julia> zero_features(g)
0

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], [[1,2.], [2,3.], [3,4.], [4,5.]]);

julia> zero_features(g)
2-element Vector{Float64}:
 0.0
 0.0

julia> g = FeatureDiGraph([1,2,3,1], [2,3,1,4], [rand(2,2) for _ in 1:4]);

julia> zero_features(g)
2×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0