Paths

Index

MultiFlows.AbstractPath

MultiFlows.VertexPath

Base.:==

Base.isempty

Base.length

Graphs.edges

Graphs.has_edge

MultiFlows.edge_indices

MultiFlows.is_path

MultiFlows.is_path

MultiFlows.path_from_edge_indices

MultiFlows.path_weight

MultiFlows.shortest_paths

Full docs

MultiFlows.AbstractPath — Type

AbstractPath

Abstract type for representing paths in a graph.

MultiFlows.VertexPath — Type

VertexPath

Concrete type for paths composed of a sequence of vertex indices.

Example

julia> g = grid((3,2))
{6, 7} undirected simple Int64 graph

julia> ds = dijkstra_shortest_paths(g, 1);

julia> p = VertexPath(enumerate_paths(ds, 6))
VertexPath{Int64}([1, 2, 5, 6])

Base.:== — Method

Base.:(==)(p1::VertexPath{T}, p2::VertexPath{T}) where {T}

Check path equality. Checks p1.vertices == p2.vertices.

Base.isempty — Method

Base.isempty(p::VertexPath)

Check if the path is empty.

Base.length — Method

Base.length(p::VertexPath; edges=false)

Returns length of the path. Default is to return number of vertices, if edges=true returns number of edges.

Example

julia> p = VertexPath([1,2,3])
VertexPath{Int64}([1, 2, 3])

julia> length(p)
3

julia> length(p, edges=true)
2

Graphs.edges — Method

Graphs.edges(p::VertexPath)

Return edge iterator of the path.

Example

julia> p = VertexPath([1,2,3])
VertexPath{Int64}([1, 2, 3])

julia> edges(p)
2-element Vector{Tuple{Int64, Int64}}:
 (1, 2)
 (2, 3)

Graphs.has_edge — Method

Graphs.has_edge(p::VertexPath{T}, s::T, t::T) where {T}

Check if the path contains edge (s,t).

Example

julia> p = VertexPath([1,2,3]);

julia> has_edge(p, 1, 2)
true

julia> has_edge(p, 1, 3)
false

MultiFlows.edge_indices — Method

edge_indices(p::VertexPath, g::AbstractGraph)

Get list of edge indices corresponding to the path.

Example

julia> p = VertexPath([1,2,5,6]);

julia> edge_indices(p, g)
3-element Vector{Int64}:
 1
 4
 7

julia> collect(edges(g))[edge_indices(p,g)]
3-element Vector{Graphs.SimpleGraphs.SimpleEdge{Int64}}:
 Edge 1 => 2
 Edge 2 => 5
 Edge 5 => 6

MultiFlows.is_path — Method

is_path(p::VertexPath, g::AbstractGraph)

Check if p is a valid path in graph g.

Example

julia> g = grid((3,2));

julia> p = VertexPath([1,2,5,6]);

julia> is_path(p, g)
true

julia> is_path(p, grid((2,1)))
false

MultiFlows.is_path — Method

is_path(p::VertexPath{T}, s::T, t::T)

Check if path is an $s-t$-path.

Example

julia> p = VertexPath([1,2,3]);

julia> is_path(p, 1, 2)
false

julia> is_path(p, 1, 3)
true

MultiFlows.path_from_edge_indices — Method

path_from_edge_indices(ei::Vector{Int64}, g::AbstractGraph)

Convert a list of edge indices to a path in graph g.

Example

julia> g = grid((3,3))
{9, 12} undirected simple Int64 graph

julia> p = VertexPath(enumerate_paths(dijkstra_shortest_paths(g, 1), 9))
VertexPath{Int64}([1, 2, 5, 6, 9])

julia> ei = edge_indices(p, g)
4-element Vector{Int64}:
  1
  4
  8
 10

julia> path_from_edge_indices(ei, g)
VertexPath{Int64}([1, 2, 5, 6, 9])

MultiFlows.path_weight — Method

path_weight(p::VertexPath, g::AbstractGraph; aggr::Function=sum, dstmx=weights(g))

Compute the weight of the path in graph g.

Example

julia> path_weight(p, g)
3

You may provide a different weight matrix

julia> w = zeros(nv(g), nv(g));

julia> w[Bool.(adjacency_matrix(g))] .= 1:2*ne(g);

julia> w
6×6 Matrix{Float64}:
 0.0  3.0  0.0  8.0   0.0   0.0
 1.0  0.0  6.0  0.0  10.0   0.0
 0.0  4.0  0.0  0.0   0.0  13.0
 2.0  0.0  0.0  0.0  11.0   0.0
 0.0  5.0  0.0  9.0   0.0  14.0
 0.0  0.0  7.0  0.0  12.0   0.0

julia> path_weight(p, g, dstmx=w)
27.0

julia> path_weight(p, g, dstmx=w, aggr=minimum)
3.0

MultiFlows.shortest_paths — Method

shortest_paths(g::AbstractGraph{T}, src::T, dst::T; K::Int64=1)

Compute shortest paths from s to t. If K=1 uses Graphs.dijkstra_shortest_paths and if K>1 uses Graphs.yen_k_shortest_paths.

Example

julia> g = grid((3,3))
{9, 12} undirected simple Int64 graph

julia> shortest_paths(g, 1, 9)
VertexPath{Int64}([1, 2, 5, 6, 9])

julia> shortest_paths(g, 1, 9, K=2)
2-element Vector{VertexPath{Int64}}:
 VertexPath{Int64}([1, 2, 5, 6, 9])
 VertexPath{Int64}([1, 4, 5, 8, 9])